chking phNo uniquevalidation-default.aspx

+++ WebApplication/App_Code/studentCls.cs
@@ -13,7 +13,7 @@ namespace WebApplication.App_Code
         public string EducationalLvl { get; set; }
         public string Skill { get; set; }
         public string Email { get; set; }
-        public int ContactNo { get; set; }
+        public double ContactNo { get; set; }
         public int QuizeId { get; set; }
     }
 }
